add database view, connect with details

+++ src/main/java/pl/edu/agh/ui/controller/InputOutputStageRefresh.java
@@ -1,8 +1,6 @@
 package pl.edu.agh.ui.controller;
 
-import pl.edu.agh.ui.util.ProcessDetailsWrapper;
-
 public interface InputOutputStageRefresh {
 
-    void refresh(Integer data);
+    void refresh(Long data);
 }
